fix(security): implement robust SSRF protection in Rust CLI
Consolidate SSRF protection into a dedicated `ssrf` module and implement
a safe `reqwest` client that validates each redirect hop against private
IP ranges and restricted hostnames.

- Create `cli/src/ssrf.rs` with `is_safe_url` and `create_safe_client_builder`.
- Update `DirectFetchProvider` and `LlmsTxtProvider` to use the safe client.
- Update `link_validator` to filter links and use the safe client.
- Centralize `is_safe_url` logic and re-export it in `cascade.rs`.
- Ensure all outbound requests in the Rust CLI are SSRF-protected.
