+++ cli/src/providers/direct_fetch.rs
@@ -3,11 +3,13 @@
 //! Basic content extraction from HTML.
 
 use crate::error::{ResolverError, detect_error_type};
+use crate::ssrf::create_safe_client_builder;
 use crate::types::ResolvedResult;
 use async_trait::async_trait;
 use std::result::Result;
 use std::sync::Arc;
 use std::sync::atomic::{AtomicBool, Ordering};
+use std::time::Duration;
 
 /// Direct HTTP fetch provider
 pub struct DirectFetchProvider {
@@ -19,7 +21,9 @@ impl DirectFetchProvider {
     /// Create a new direct fetch provider
     pub fn new() -> Self {
         Self {
-            client: reqwest::Client::new(),
+            client: create_safe_client_builder(Duration::from_secs(10))
+                .build()
+                .unwrap_or_default(),
             rate_limited: Arc::new(AtomicBool::new(false)),
         }
     }
